refactor currently testing

+++ lib/nu_system.sh
@@ -81,22 +81,29 @@ require_tmp () {
 }
 
 retire_tmp () {
+	local opt_keep=
+	while getopts k OPT; do case $OPT in
+		k) opt_keep=y;;
+	esac; shift; done
+	
 	[ $# = 1 ]
 	[ -n "$1" ]
 	
-	if [ -n "${OPT_DEBUG-}" ]; then
-		require_tmp -d -l debug_out _retire_tmp_debug_out
-		if eval [ -e \"\$_retire_tmp_debug_out\/\$1\" ]; then
-			eval mv -n \"\$_retire_tmp_debug_out\/\$1\" \"\$_retire_tmp_debug_out\/0.\$1\"
+	if [ -z "$opt_keep" ]; then
+		if [ -n "${OPT_DEBUG-}" ]; then
+			require_tmp -d -l debug_out _retire_tmp_debug_out
+			if eval [ -e \"\$_retire_tmp_debug_out\/\$1\" ]; then
+				eval mv -n \"\$_retire_tmp_debug_out\/\$1\" \"\$_retire_tmp_debug_out\/0.\$1\"
+			fi
+			local i; unset i
+			while eval [ -e \"\$_retire_tmp_debug_out\/${i:-0}.\$1\" ]; do
+				: ${i:=0}
+				i=$(($i+1))
+			done
+			eval mv -n \"\$$1\" \"\$_retire_tmp_debug_out\/${i:+$i.}\$1\"
+		else
+			eval rm -r \"\$$1\"
 		fi
-		local i
-		while eval [ -e \"\$_retire_tmp_debug_out\/${i:-0}.\$1\" ]; do
-			: ${i:=0}
-			i=$(($i+1))
-		done
-		eval mv -n \"\$$1\" \"\$_retire_tmp_debug_out\/${i:+$i.}\$1\"
-	else
-		eval rm -r \"\$$1\"
 	fi
 	eval unset "$1"
 }
@@ -108,28 +115,36 @@ choose_random () {
 }
 
 sets_union () {
-	local ret_tmp=$1; shift
+	local ret_var=$1; shift
 	
 	[ $# -ge 1 ]
-	[ -w "$ret_tmp" -a ! -s "$ret_tmp" ]
+	
+	local ret_tmp=
+	require_tmp -l $ret_var ret_tmp
 	
 	cat "$@" | sort -u >| "$ret_tmp"
+	setvar $ret_var "$ret_tmp"
 }
 
 sets_sym_diff () {
-	local ret_tmp=$1; shift
+	local ret_var=$1; shift
 	
 	[ $# = 2 ]
-	[ -w "$ret_tmp" -a ! -s "$ret_tmp" ]
+	
+	local ret_tmp=
+	require_tmp -l $ret_var ret_tmp
 	
 	cat "$@" | sort | uniq -u >| "$ret_tmp"
+	setvar $ret_var "$ret_tmp"
 }
 
 sets_intrsctn () {
-	local ret_tmp=$1; shift
+	local ret_var=$1; shift
 	
 	[ $# -ge 2 ]
-	[ -w "$ret_tmp" -a ! -s "$ret_tmp" ]
+	
+	local ret_tmp=
+	require_tmp -l $ret_var ret_tmp
 	
 	case $# in
 		2)
@@ -138,4 +153,5 @@ sets_intrsctn () {
 		*)
 			cat "$@" | sort | uniq -c | sed -nEe "/^[[:blank:]]*$# /{s///;p;}" >| "$ret_tmp"
 	esac
+	setvar $ret_var "$ret_tmp"
 }
